gomod: fetch_gomod_source: Rename env_vars -> env_vars_template
[COSMETIC CHANGE]
Make it absolutely clear these are variables we export for the hermetic
build to consume and not the ones we use for our own purposes. This is
because we're going to move another set of env variables here in future
patches so we need to distinguish between them.

Take the opportunity to also move the variable set definition closer to
where they're used.

Signed-off-by: Erik Skultety <eskultet@redhat.com>