Replace mentions of the old containerbuildsystem organization
The project moved to a new GitHub organization "hermetoproject". Change
the references accordingly with the exception of tests where some of
them would start failing due to references to tags only present in the
old organization. That is material for a follow-up work so this patch
is best effort only for now.

Signed-off-by: Erik Skultety <eskultet@redhat.com>