yarn (classic) - Verify repository before resolving a yarn package
Currently, a user only specify 'yarn' in the CLI as a requested package
manager, regarding of the major version of yarn. We check yarn.lock for
a specific pattern if the project uses yarn classic (v1) or yarn berry
(v3, v4).

Let's verify the repository before we start with resolving the project
and fetching the dependencies. This will remove duplicate logs:

"Fetching the yarn dependencies at ..."

that are exposed to a user if `fetch_yarn_classic_source` throws and
exception and then we start resolving the project using yarn berry
(v3, v4).

Also, it will remove empty `yarn-classic` directory that would have been
created when processing yarn classic package.

To reproduce it, just run `cachi2 feth-deps yarn` inside:
https://github.com/hermetoproject/integration-tests/tree/yarn/e2e

Signed-off-by: Michal Šoltis <msoltis@redhat.com>