Use temporary directory when resolving each package manager
Closes https://github.com/hermetoproject/cachi2/issues/707
Closes https://github.com/hermetoproject/cachi2/issues/712

All package managers (except yarn) work on top of the input source
directory directly which may result in dirtying the repo which is
completely undesirable.

The changes simplify the package resolution logic by removing the
conditional check for yarn_packages. This ensures that the temporary
directory is always used, making the behavior consistent across all
package types. The test suite has also been updated to reflect this
change.

This refactor aligns with the goal of maintaining a uniform approach to
handling source directories, regardless of the package manager being
used.

Signed-off-by: Michal Šoltis <msoltis@redhat.com>