models: property_semantics: Use enum instead of Literal for PropertyName
Literal isn't the best choice due to its inflexibility when it comes to
both comparing values against it which requires a harcoded string
literal on one side of the comparison as well as its inability to use
additional string formatting.

Future patches will rebrand the project to 'hermeto' from 'cachi2'
which also means we need to start reflect that in the SBOM labels we
create. However, plain string replacement would lead to breaking
backwards compatibility, i.e. if one would use 'cachi2' as the
entrypoint they'd still see e.g. 'hermeto:found_by' in the generated
SBOM - we don't want that. Moreover, Enums have runtime validations
which pydantic respects and so we don't need a dedicated field
validator for the Property model, see the following error:

    pydantic_core._pydantic_core.ValidationError: 1 validation error for
    Property name
        Input should be 'cachi2:bundler:package:binary',
                        'cachi2:found_by',
                        ...
                        or 'cdx:npm:package:development'
        [type=enum, input_value='dsfsdaf', input_type=str]

and so we don't need to be worried that replacing Literal with Enum
would inherently open us up to sneaking OR not validating incorrect
values passed to property names.

Therefore, this patch replaces the PropertyName Literal structure with
a string Enum.

Signed-off-by: Erik Skultety <eskultet@redhat.com>