tests: unit: Drop isodate fixture - use more targeted mocking
There are 2 motivators behind this patch:
1) any attempt to put a debugger breakpoint to any of the SPDX-related
   SBOM unit tests automatically ends with the following error:

   "sqlite3.ProgrammingError: Error binding parameter 1: type 'MagicMock'
        is not supported."

It is unclear what exactly is using sqlite3, most likely it's
IPython for its history (ipdb uses ipython underneath), it's none of
our dependencies. That said, the error stems from the fact that sqlite3
doesn't have a MagicMock adapter which we're making use of currently by
the nature of the mocking 'isodate' fixture.
The main problem here though is how what isodate actually mocks.
Firstly, it mocks the whole datetime module and forces a formatstring
that only SPDX needs onto potentially any caller who'd wish to mock
datetime. Additionally, it only mocks the strftime function and nothing
else. Datetime is a harmless module which can easily use if we pass
if we work with a datetime object of a particular timestamp - epoch
time in this case for simplicity.

2) Repetitive hardcoded time values in the test cases instead of using
a constant

Static strings are error-prone to typos and any kind of top-level
changes, covering this with a constant is safer and easier to maintain.

Based on the reasoning above, this patch replaces and drops the current
isodate mocking fixture and replaces it with a targeted spdx_now() mock
in modules affected by the function in question.

Signed-off-by: Erik Skultety <eskultet@redhat.com>