core: resolver: Fix how project files paths are updated
We have essentially 2 scenarios we need to cover:
1) Bundler
    - we do need to override user configuration with a file, but unlike
      other backends (i.e. pip) we can put the override file to our
      output dir and reference it from there
    - we must not try updating the project file path because it's
      already properly resolved

2) Pip/NPM
    - these require changes to project files directly inside the source
      repository and so we must update the paths correctly
    - since our default working directory is now temporary, the project
      files paths' are resolved within that and we need to fix them
    - since the default working directory is inside the source
      repository, it's easy because the project file template is always
      going to be a relative subpath of the other and we can safely
      replace

The problem appears when the stars align and the output dir, the source
dir, and CWD are all identical (yes, this happened). With our current
logic we'd process Bundler just fine, because that one places its
project file override to the output dir directly and our temporary
working copy (i.e. <tmp>.cachi2-source-copy) directory isn't a subpath
of it. However, Pip/NPM backends need to place their overrides back to
the original source repo and that's where we'd cause a hermetic build
to fail because the temporary working copy is a subpath of the absolute
path for the project file residing in the original source directory and
so we wouldn't update because it's "seemingly" already there. By
forcing and incorrect path for the project file override the subsequent
hermetic build would fail because the file wouldn't be in its expected
location either by Pip or NPM and either backend would instead try to
connect to the internet.

This patch forces the replacement of all paths. If the replacement
fails we skip such path. This should cover all cases because our
temporary working copy is always within the original source path with
case 3 being the only determinant (at this point in time) of the
outcome.
This patch also re-enables the resolver project path mangling unit test
since now the issue is gone and the test verifies it.

Fixes: d4d8dd9ad3878fb963b5ab97faf1168ea1520d03

Signed-off-by: Erik Skultety <eskultet@redhat.com>
(cherry picked from commit 6841ffcedbe11ad96a650591367abb96ef9bc8e8)
Signed-off-by: Erik Skultety <eskultet@redhat.com>