noxfile: Use --no-deps with pip requirements installation
The upstream CI fails with:

    ERROR: In --require-hashes mode, all requirements must have their
    versions pinned with ==. These do not: coverage[toml]>=7.5

By default, pip tries to resolve transitive dependencies of packages
when reading the requirements files. Well, that fails if the
'require-hashes' mode is ON. In case one uses a fully resolved and
pinned requirements file (we do), pip shouldn't need to spend time
trying to re-resolve something that's already resolved, it should use
it directly.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Suggested-by: Alexey Ovchinnikov <aovchinn@redhat.com>