gomod: Run the setup toolchain code in fetch_gomod_source
In order to ensure consistency of the go toolchain version that was
used across the whole package processing sequence, we have to run the
toolchain setup code much earlier, i.e. not in _resolve_gomod, but in
fetch_gomod_source. This patch uses a simple and straightforward
approach to make unit tests pass, but proper logic will be introduced
in a future patch that will replace this function with a different
selector logic.

Signed-off-by: Erik Skultety <eskultet@redhat.com>