gomod: Go: Introduce a replacement for _install helper
Introduce a new convenience class method that conveys the message
better through a new interface. This helper is based on the existing
'Go._install' helper and so it's almost completely identical, but there
are tiny differences between the two, but the idea remains the same -
install a matching Go SDK from the official sources.
Since there's only a couple of places where we need it, make use of it
in this patch as well.

Signed-off-by: Erik Skultety <eskultet@redhat.com>