chore: Improve messaging for corrupt Cargo.locks
When a Cargo.lock and Cargo.toml do not match `cargo vendor --locked`
will refuse to proceed. This is a corner case which manifests itself
in several Python packages extended with Rust. While making sure that
locks match to configs is not a direct responsibility of Hermeto
the messaging around this specific failure was confusing.

This change adds a dedicated exception class to raise when there is a
mismatch which contains a more thorough explanation of the problem.

Signed-off-by: Alexey Ovchinnikov <aovchinn@redhat.com>