cargo: Regenerate lock files in permissive mode
Previously if cargo found a lock file that did not match project
configuration it would fail the entire prefetch. This was found
to affect certain Rust-based Python dependencies, (jiter and ruff,
likely others as well). This change make cargo PM check if the mode is
permissive and if yes then try to regenerate Cargo.lock first and
attempt vendoring the package again.

Signed-off-by: Alexey Ovchinnikov <aovchinn@redhat.com>