cargo: Force old-style config to the new style
Turns out it is still possible to occasionally find projects that
keep pre-v1.39.0 style config which with any newer cargo would
take precedence over any existing new style config. Since the only
important difference between the two styles of configs is their name
just renaming an old style one is enough to proceed (unless the
project is to be built with a very old Cargo+Rust combination, but
this does not look likely at the moment).

This commit adds a transparent guard to implicitly rename the
config and to thus ensure that Hermeto's changes to it are respected.

Signed-off-by: Alexey Ovchinnikov <aovchinn@redhat.com>