logging: Introduce a new logger adapter
In order to selectively report a warning rather than an error in cases
where we wish to apply the effect of the --mode=permissive CLI flag we
can either introduce a bunch of if-else clauses in all places where
applicable or we can instead opt for an idiomatic solution involving
logger adapters [1] which allow contextual changes to the data being
reported and how they're logged. The advantage of a logger adapter is
that it's essentially a wrapper over a regular logger instance and so
it implements the same public interface. That's perfect for our cause
because we can define a new method that would contextually report a
warning or an error in scenarios where the CLI flag should take effect.
All other places would continue functioning identically via existing
logger invocations thanks to the public interface compatibility.

[1] https://docs.python.org/3/howto/logging-cookbook.html#using-loggeradapters-to-impart-contextual-information

Signed-off-by: Erik Skultety <eskultet@redhat.com>