package_managers: gomod: Apply the permissive mode on vendoring checks
One of the requests we've received on several occasions is that our
strict checking of the vendor/ directory content is preventing our
users from building projects successfully in network-isolated
environments, because they may be employing kludges and quirks that
happen to have side-effects in the vendor directory and our git
detection logic would flag those.

With the contextual enforcing mode logging adapter in place, converting
the vendor changes related checks to mere warnings is pretty simple.
The only suboptimal thing that needs to be resolved as part of a bigger
overhaul is avoid passing the enforcing mode setting around, especially
in nested functions, that just clutters the function signatures with a
setting that has a global effect and so should be taken care of in a
different manner (e.g. a Singleton).

Resolves https://github.com/hermetoproject/hermeto/issues/933

Signed-off-by: Erik Skultety <eskultet@redhat.com>