tests: unit: gomod: test_vendor_changed: Apply pytest IDs to variants
When working on some vendor impacting changes I naturally faced a
bunch of test failures in test_vendor_changed. Restricting the test run
to just a single failing test variant with a debugger connected is much
easier when the test case variants are named in human readable form -->
apply test IDs to the variants.

Signed-off-by: Erik Skultety <eskultet@redhat.com>