tests: integration: Add a new TestParameters field - global_flags
Currently, flags are only meant for the given sub-command (fetch-deps
in most cases). Typer doesn't allow global flags to be specified
arbitrarily on the CLI and so we need to add a new argument to the
TestParameters class to enable passing of the '--mode=strict|permissive'
flag.

Use the opportunity to drop a redundant test_params.flags boolean check
since there's a default factory defined for the field and so the worst
case scenario would be adding an empty list.

The effect of this will be visible in a future patch exercising the
'--mode=permissive' CLI option.

Signed-off-by: Erik Skultety <eskultet@redhat.com>