tests: integration: Flip most rpm tests to CentOS Stream 9
Our current Fedora 39-based on is failing because 39 has been EOL'd
for a while and so under Fedora's policy has moved to archive [1], [2].

[1] https://dl.fedoraproject.org/pub/fedora/linux/releases/39/
[2] https://dl.fedoraproject.org/pub/archive/fedora/linux/releases/39/

This is inevitably going to happen with future EOL'd releases again.
UBI/AlmaLinux/RockyLinux aren't an option because they only provide
the latest build of any package so our static URLs might break
often. CentOS Stream seems to follow a different policy and provides up
to 5 package builds [1].

Since the CentOS Stream images come with empty RPM GPG database (but
with GPG keys pre-installed) the Containerfile had to be adjusted so
that we don't run into GPG verification issues during the test.

[1] https://pkgs.org/search/

Signed-off-by: Erik Skultety <eskultet@redhat.com>