Add RPM modularity label
Add a new property to capture RPM modularity labels for enhanced package
identification. This is necessary because modularity labels provide
critical metadata for packages in modular repositories.

Modularity label is guaranteed to be in the RPM tags for RPM modules.
Providing this additional information about an RPM enhances accuracy
of our SBOMs.

Also, it helps during static analysis on checking if a container image
matches with a produced SBOM. Previously no modularity information has
been present in our SBOMs.

Signed-off-by: Michal Šoltis <msoltis@redhat.com>