Add requirements-build.txt file
The setuptools-scm library is needed as a dependency when building the
project. The lack of this dependency will make the build fail in a
hermetic environment.

The setuptools library is also included here since it is a build
dependency as well, although that one is currently also included in the
requirements.txt file as a transitive dependency.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>