integration tests: Boilerplate for choosing the container engine
This patch moves the `run_cmd` function from the integration tests utils
to its own separate file and creates the boilerplate for allowing the
container engine to be chosen.

For now, only the PodmanEngine exists, and no mechanism to actually
choose the engine was implement. This will be done in follow-up
commits.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>