Option: Make entrypoint an explicit parameter
Considering buildah doesn't have a flag for entrypoint, making it a
separate function parameter in the test helpers would avoid the long
logic of handling corner cases of how --entrypoint can be passed.

This is just a PoC commit, let me know if you like the direction so I
can refactor the code properly.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>