toolbox: Containerfile: Add a host-runner wrapper script
Running some tooling relying on isolation, podman included, is either
not possible or extremely limited within OCI containers. Toolbox
currently doesn't have a native way of dealing with this problem on a
generic level, but it provides a 'flatpak-spawn' command which can
offload a command to the container host, which is exactly what the
toolbox binary inside a toolbox container will do.
Similarly, we can provide a thin wrapper script for the most common
tooling in the small hermeto world (podman, buildah) that we'd create
symlinks to the helper wrapper and let the host deal with the
execution (until toolbox comes up with native support for this [1]).
This patch was inspired by [2].

[1] https://github.com/containers/toolbox/issues/145
[2] https://github.com/containers/toolbox/issues/145#issuecomment-582040463

Signed-off-by: Erik Skultety <eskultet@redhat.com>