models: sbom: Placate type checker
'Argument of type "Any | None" cannot be assigned to parameter "purl"
of type "str"'

was reported by pyright and rightfully so, our Component model mandates
both name and purl to be defined. We even have a model correctness unit
test specifcally covering this scenario, so it's impossible to get None
here. We could access the dict directly, but let's be correct AND
conservative at the same time and instead of None, let's use an empty
string as a default.

Use the opportunity to drop the explicit None default from the version
extractor since None is always the default with dict.get().

Signed-off-by: Erik Skultety <eskultet@redhat.com>