integration tests: Add a parameter for defining the container entrypoint
Previously, the `--entrypoint` flag was simply passed to the test
containers as part of the generic `podman_flags` parameter. This works
well for Podman, but other container engines might have different
approaches to deal with the entrypoint (e.g. Buildah).

This commit aims to simplify the implementation of future container
engines.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>