Dockerfile: ship both cargo & rustc
The "cargo" executable alone may not be enough to resolve/check Rust
dependencies. In some cases, it tries to run "rustc", thus failing
because it is missing in the container image.

To make sure that "cargo" can run properly all the time, install that
and "rust" directly in the base image used for the final image, rather
than copying the "cargo" executable from an helper image. This way not
only "cargo" is available, a working "rustc" is too.

Resolves: https://github.com/hermetoproject/hermeto/issues/1061

Fixes commit 55c218dc529e99b284a66d04ef48aa56d3fcc4c0.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>