tests: integration: Replace forced subprocess options with defaults
Now that we have support for more than a single container engine
powering the execution of integration tests it's time to evaluate our
decision to force some subprocess options. The developers are here in
full charge of how and with what arguments the integration tests are
invoked and so a handful of default options that can be overridden in
certain cases rather than doing it vice versa, i.e. by providing
options with test invocation and overriding those particular calls with
some forced defaults, change it and allow individual container engine
instances to override the subprocess options at their will. The effect
of this patch will be more visible in a future patch.

Signed-off-by: Erik Skultety <eskultet@redhat.com>