tests: integration: conftest: Add a new fixture netrc_file
The motivation behind this is to drop the conditional execution of a
different 'run_cmd_on_image' code path that depends on whether
HERMETO_TEST_NETRC_CONTENT was provided. The thing is, our noxfile
always sets the variable for integration tests which means that what
used to be an optional code execution of our integration tests is now
the main execution path. Fix this problem by extracting the netrc
content handling to a global fixture marked as autouse.

Now, the conditional execution may look completely harmless because
there's no other code in the method at hand but that will change in
future patches where this issue would be much more noticeable.

By using the fixture we can now simply pass one more global mount point
and drop the questionable code path in 'run_cmd_on_image'.

Signed-off-by: Erik Skultety <eskultet@redhat.com>