gomod: Record the Go instance in the context manager
We want to make sure we use the same Go toolchain to cleanup any
temporary files as the one that was used to process the input package.

The problem at hand is that in order to customize a subclassed context
manager we now have to introduce a custom __init__ method with a custom
private attribute which also means the original type hint '[str]' is
no longer true and so the fetch_gomod_source caller has to be adjusted
as well.

Signed-off-by: Erik Skultety <eskultet@redhat.com>