cli: do not error out on long fetch-deps inputs
As a result of a recently introduced possibility to use a filename/path
for the fetch-deps command, the input is first tried as filename, and in
case it is not then it is loaded as JSON.

Now, in case the input string is very long, then stat'ing the string
results in an OSError exception, which causes the whole application to
stack trace. Instead, silently accept that failure and fallback to JSON
as previously done.

Fixes commit 01140a3f034dabd5e97fac631a8f8ceaf0bd416d

Signed-off-by: Pino Toscano <ptoscano@redhat.com>