gomod: GoWork: Add a convenience '@rooted_path' property
Mostly a cosmetic change, but GoWork().path is invoked the premise
should be that we'd return a path object which is much easier to work
with than rooted path, rooted path really only is for validation, but
not something convenient to work with.

At the same time, adjust the existing @path property so that it no
longer returns a rooted path instance. Also, since it's now computed,
make it a cached property.

Signed-off-by: Erik Skultety <eskultet@redhat.com>