gomod: Go: Make the class immutable
Future patches will add logic that will query installed toolchains in
the system and so we need to make sure that the class is hashable and
we can insert it into a set.
The standard way to make a class immutable in Python seems to be making
it a frozen dataclass that already provides most of the bits we need.
We only need to add one of the ordering dunder methods (__lt__ in this
case) to support total ordering in future patches which will select the
correct toolchain version depending on the criteria.

Whatever extra handling we'd ever need to do during __init__ has to go
to __post_init__ (dataclass's approach) otherwise we'd not be able to
compute dynamic values we don't know upfront. Introducing the
__post_init__ constructor just to keep this in mind in the future.

Signed-off-by: Erik Skultety <eskultet@redhat.com>