gomod: Go: Make the class immutable
Future patches will add logic that will query installed toolchains in
the system and so we need to make sure that the class is hashable and
we can insert it into a set.
The standard way to make a class immutable in Python seems to be making
it a frozen dataclass that already provides most of the bits we need.

We also apply the `total_ordering` decorator so that we only need to
define one of the ordering dunder methods (__lt__ in this case) to
support total ordering for future patches which will select the correct
toolchain version depending on the criteria.

Whatever extra handling we'd ever need to do during __init__ has to go
to __post_init__ (dataclass's approach) otherwise we'd not be able to
compute dynamic values we don't know upfront. In our case it's
primarily about making sure that:
    1. any provided string literal resolves to an abs path to a binary
    2. our own default string value 'go' resolves to an abs path

Since we're now enforcing an absolute path for the 'binary' field
(either implicit or explicit), a few unit tests have to adjusted to
assert a path constant, not the 'go' string literal.

Signed-off-by: Erik Skultety <eskultet@redhat.com>