gomod: Go: Use absolute path as the default for the binary
Doing this beforehand will reduce the optical noise in subsequent
patches that make further transformation to the class on the quest of
making it immutable.
Note that the default path we chose for the Go toolchain points to
/usr/local/bin/go which is one of our locations inside the container
image.

Since the default is a standard system path this also enables us to
make use of the GO_CMD_PATH variable in tests more.

THIS PATH TEMPORARILY BREAKS NON-CONTAINERIZED USAGE OF HERMETO WITH GO
PREFETCHING.

^This is to keep the code base bisectable going forward and a future
patch will lift this limitation when the toolchain version selector
logic is fully refactored.

Signed-off-by: Erik Skultety <eskultet@redhat.com>