tests: unit: gomod: Point GO_CMD_PATH to /usr/local/bin/go
This is the path we install our custom toolchain versions in the
container image, so let's make it matching.
For unit tests, the path doesn't matter much, but it'll start to matter
in the next patch where we use the same path as the default for the Go
toolchain wrapping class where working unit tests and integration tests
is important.

Signed-off-by: Erik Skultety <eskultet@redhat.com>