pip: Remove binary filters from requirements validation
Previously, if a user wanted to download a wheel as an URL dependency,
he needed to specifiy boolean `allow_binary` flag to enable processing
wheels.

This behavior does not make sense anymore, since we adopted more complex
binary options and deprecated the boolean flag.

The wheel is still reported in the SBOM, so no change in that regard.

Signed-off-by: Michal Šoltis <msoltis@redhat.com>