bundler: Add platform filtering logic
Add new `GemsFilter` class that handles platform filtering for gem
dependencies. The filter can handle four different scenarios based on
whether specific packages or platforms are requested.

When no filters are set, it prefers binary packages by removing the
ruby platform when other platforms are available. When specific
platforms are requested, it limits gems to those platforms. The filter
makes sure gems not in the package list stick to the ruby platform.

More information about the proposed approach is in the design document:
https://github.com/hermetoproject/hermeto/blob/main/docs/design/binary-platform-filtering.md

Signed-off-by: Michal Šoltis <msoltis@redhat.com>